@@ -1,4 +1,4 @@
-name: vians
+name: cnab
 description: Find your patner with us
 authors:
   - kevin ASIEDU<kevinasiedu.contact@gmail.com>
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -50,6 +50,9 @@ dependencies:
   pin_input_text_field: ^4.1.1
   flutter_svg: ^0.22.0
   image_picker: ^0.8.4+1
+  change_app_package_name: ^1.1.0
+  shared_preferences: ^2.0.7
+  fluttertoast: ^8.0.9
 
 
 dev_dependencies:
