@@ -1,11 +1,10 @@
 /* This is country list class */
 class Country {
-   String name;
-   String code;
-   String flag;
+  String? name;
+  String? code;
+  String? flag;
 
   Country(this.name, this.code, this.flag);
-
 
   // Simple constructor
   Country.simple() {
@@ -16,9 +15,8 @@ class Country {
 
   // Country from json
   Country.fromJson(Map<dynamic, dynamic> data) {
-   this.name = data['text'];
-   this.flag = data['map'];
-   this.code = data['callingCodes'];
+    this.name = data['text'];
+    this.flag = data['map'];
+    this.code = data['callingCodes'];
   }
-
 }
