@@ -1,11 +1,10 @@
+import 'package:cnab/screens/splashScreen.dart';
+import 'package:cnab/translations.dart';
+import 'package:cnab/utils/design.util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:vians/screens/splashScreen.dart';
-import 'package:vians/translations.dart';
-import 'package:vians/utils/design.util.dart';
 
 class MyApp extends StatefulWidget {
   _MyAppState createState() => _MyAppState();
@@ -18,24 +17,16 @@ class _MyAppState extends State<MyApp> {
     //     statusBarColor: Colors.white
     // ));
     return MaterialApp(
-      title: 'dope',
+      title: 'Cnab',
       debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primaryColor: primaryColor,
-        textTheme: GoogleFonts.poppinsTextTheme(
-          Theme.of(context).textTheme
-        )
-      ),
+      theme: ThemeData(primaryColor: primaryColor, textTheme: GoogleFonts.poppinsTextTheme(Theme.of(context).textTheme)),
       localizationsDelegates: [
         const TranslationsDelegate(),
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate,
         GlobalCupertinoLocalizations.delegate,
       ],
-      supportedLocales: [
-        const Locale('en', ''),
-        const Locale('fr', '')
-      ],
+      supportedLocales: [const Locale('en', ''), const Locale('fr', '')],
       home: SplashScreen(),
     );
   }
