@@ -1,7 +1,10 @@
+import 'package:cnab/services/storage_service.dart';
 import 'package:flutter/material.dart';
 
 import 'app.dart';
 
-void main() {
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await StorageService().init();
   runApp(MyApp());
 }
